Feat/oauth meinbildungsraum (#155)
* Add and configure mozilla-django-oidc

The views that are provided by this library are not yet used. We need to
expose them to the badgr-ui, make the relevant requests in the ui and
probably create an interface in the UI that allows for adding an E-Mail
address to the account that is authenticated via OIDC. For this also
further configurations to the library have to be made, because currently
it will try to automatically connect OIDC accounts to django users via
the E-Mail address.

* Add view for trying OIDC

* Adjust scopes and add endpoints

* Add custom auth backend

* Relate OIDC users to badgr users and don't create duplicates

* Add DRF authentication classes

* Remove SessionRefresh because of CORS error

* WIP

* WIP create access token

* Access token generation

* Attach email to OIDC user

* Log out from session

Logging out from the session after 60s or after
the access token was obtained. Also, don't store
the OIDC token in the session.

* Add login base url to allow token validation

* Disale OIDC debug view

This commit disables the access to the OIDC debug view from the backend.

* Adjust settings

Use PKCE and allow credentials in global settings.

* Add mbr_user property

Users now have the `mbr_user` property (default `False`), indicating
whether this user was automatically created by the OIDC flow. This will
also be used to mark users that should get logged out of mbr on logout
of oeb.

* Also logout from mbr on oeb logout

If the user logged in via mbr and they logout via oeb, this (together
with the changes in badgr-ui) will redirect them to a logout page of
mbr, where they can confirm to also logout from mbr and then to go back
to oeb.

* Remove mbr_user property

The property is not needed after all, because whether the user logged in
via mbr is now stored in the frontend

* Move logout redirect endpoint to oidc view and add triggered logout endpoint

* Store session ID alongside access token

This adds a one to one table storing the session ID belonging to access
tokens created after OIDC authentication. This allows for a triggered
logout via OIDC, since the LOGOUT_TOKEN also contains this session ID.
The creation is tested, the logout can't be tested yet though, because
keycloak first has to be configured to also use our endpoint.

* Add TODO

* Custom claims verification to avoid warning message and ensure sub is contained in claims

* Trying out OIDC access token authentication

* Remove accesstokensessionid and triggered logout

Those functionalities aren't required anymore with the new approach of
using OIDC access tokens instead of our own ones.

* Cleanup OIDC access token implementation

* Document OIDC in README

* Use prompt=none query parameter for silent token renewal

* WIP Copied refresh token implementation

* Cleanup and minor fixes

* Fix a few more minor bugs

* A few fixes and a realization

* Functional refresh token implementation

* Fix PKCE & minor cleanup

Re-added PKCE support to the custom authenticate method. Also removed
some TODOs and the non-needed middleware.

* Remove uneeded code & TODO

* Correct typo & remove uneeded stuff

* fix: typo

---------

Co-authored-by: Timber <timber@opensenselab.org>