Feat/change email of institution on ownership change (#182)
* fix: infinite recursion

Remove the storage of the original information in __init__, since this
caused infinite recursion on deletion of users associated with an
institution. Instead in saving I receive the old version from the
database to compare for changes.

* feat: transfer ownership on user deletion

If a user is deleted the ownership of all associated issuers is
transferred to the next staff with the highest privileges. Also the
contact email of the issuer is set to the primary email of the new
owner. "Owner" in this sense also implicates creator, to ensure that no
issuer exists without a creator (even though this is kind of a lie).

* feat: also adjust mail address if owner removed

So far the changes were only regarding the owner getting *deleted*, not
*removed*. With these changes the contact mail address is also adjusted
if the user with the current contact mail is removed from the
institution. This was basically already implemented, the delete method
just had to be called one by one, the creator / owner relationship had
to be adjusted a bit in the owner retaining method and some cache issues
had to be fixed.

* fix: Attach creator as owner for no other member

If no other member exists for the issuer, the creator ir (re)-attached
as owner of the issuer. This is also the case for the creation of the
issuer.

* fix: also allow removal of creator

Before the creator also had to be assigned as staff member; now the user
only needs to *exist*. This was an error of thinking, because I already
intended it to be this way.

---------

Co-authored-by: Timber <timber@opensenselab.org>
Co-authored-by: Yousef Qamaz <40634685+YouQam@users.noreply.github.com>